+++ spec/transpec/context_spec.rb
@@ -106,6 +106,11 @@ module Transpec
       end
 
       def eval_with_rspec_in_context(eval_source)
+        # Clear SPEC_OPTS environment variable so that this spec does not fail
+        # with dynamic analysis in self-testing.
+        original_spec_opts = ENV['SPEC_OPTS']
+        ENV['SPEC_OPTS'] = nil
+
         result_path = 'result'
 
         helper_source = <<-END
@@ -117,11 +122,14 @@ module Transpec
         END
 
         source_path = 'spec.rb'
+
         File.write(source_path, helper_source + source)
 
         `rspec #{source_path}`
 
         Marshal.load(File.read(result_path))
+      ensure
+        ENV['SPEC_OPTS'] = original_spec_opts
       end
 
       describe '#non_monkey_patch_expectation_available?' do